src/requirements.txt
requests
pandas
numpy
Pillow
seaborn
matplotlib
scipy
scikit-learn
xgboost
sentence-transformers
tqdm